requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
python-docx>=0.8.11
PyPDF2>=3.0.0
python-pptx>=0.6.21
openai>=1.0.0
langchain>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0